Feedback fixes, update changes file
