+++ lucene/CHANGES.txt
@@ -75,9 +75,9 @@ API Changes
 * LUCENE-10368: IntTaxonomyFacets has been deprecated and is no longer a supported extension point
   for user-created faceting implementations. (Greg Miller)
 
-* LUCENE-10050: Deprecate DrillSideways#search(Query, Collector) in favor of DrillSideways#search(Query, CollectorManager).
-  This reflects the change (LUCENE-10002) being made in IndexSearcher#search that trends towards using CollectorManagers over
-  Collectors. (Gautam Worah)
+* LUCENE-10050: Deprecate DrillSideways#search(Query, Collector) in favor of
+  DrillSideways#search(Query, CollectorManager). This reflects the change (LUCENE-10002) being made in
+  IndexSearcher#search that trends towards using CollectorManagers over Collectors. (Gautam Worah)
 
 New Features
 ---------------------
